remove duplicated `Signature` code
