+++ src/lib.rs
@@ -63,3 +63,4 @@ pub mod multi_signature;
 pub mod signature_2018;
 pub mod pok_sig_2018;
 pub mod rsssig;
+pub mod message_structure;
