+++ src/message_structure/mod.rs
@@ -1,0 +1,1 @@
+pub mod did;
